src/solokit/init/dependency_installer.py
"""
Dependency Installer Module

Handles dependency installation using exact versions from stack-versions.yaml.
"""

from __future__ import annotations

import logging
import sys
from pathlib import Path
from typing import Any, Literal, cast

import yaml

from solokit.core.command_runner import CommandRunner
from solokit.core.exceptions import CommandExecutionError, FileOperationError

logger = logging.getLogger(__name__)


def load_stack_versions() -> dict[str, Any]:
    """
    Load validated versions from stack-versions.yaml.

    Returns:
        Stack versions dictionary

    Raises:
        FileOperationError: If file not found or invalid YAML
    """
    versions_file = Path(__file__).parent.parent / "templates" / "stack-versions.yaml"

    if not versions_file.exists():
        raise FileOperationError(
            operation="load",
            file_path=str(versions_file),
            details="stack-versions.yaml not found",
        )

    try:
        with open(versions_file) as f:
            return cast(dict[str, Any], yaml.safe_load(f))
    except yaml.YAMLError as e:
        raise FileOperationError(
            operation="parse",
            file_path=str(versions_file),
            details=f"Invalid YAML: {str(e)}",
            cause=e,
        )


def get_installation_commands(
    stack_id: str,
    tier: str,
) -> dict[str, Any]:
    """
    Get installation commands for a stack and tier.

    Args:
        stack_id: Stack identifier (e.g., "saas_t3")
        tier: Tier level (e.g., "tier-1")

    Returns:
        Installation commands dictionary from stack-versions.yaml

    Raises:
        FileOperationError: If stack or tier not found
    """
    versions = load_stack_versions()

    if stack_id not in versions["stacks"]:
        raise FileOperationError(
            operation="lookup",
            file_path="stack-versions.yaml",
            details=f"Stack '{stack_id}' not found in versions file",
        )

    stack = versions["stacks"][stack_id]

    if "installation" not in stack:
        raise FileOperationError(
            operation="lookup",
            file_path="stack-versions.yaml",
            details=f"No installation commands found for stack '{stack_id}'",
        )

    return cast(dict[str, Any], stack["installation"])


def install_npm_dependencies(
    template_id: str,
    tier: Literal[
        "tier-1-essential", "tier-2-standard", "tier-3-comprehensive", "tier-4-production"
    ],
    project_root: Path | None = None,
) -> bool:
    """
    Install npm dependencies for a Next.js-based stack.

    Args:
        template_id: Template identifier (e.g., "saas_t3")
        tier: Quality tier
        project_root: Project root directory

    Returns:
        True if installation succeeded

    Raises:
        CommandExecutionError: If npm installation fails critically
    """
    if project_root is None:
        project_root = Path.cwd()

    # Map template_id to stack_id
    template_to_stack = {
        "saas_t3": "saas_t3",
        "dashboard_refine": "dashboard_refine",
        "fullstack_nextjs": "fullstack_nextjs",
    }

    stack_id = template_to_stack.get(template_id)
    if not stack_id:
        logger.warning(f"Unknown template ID: {template_id}")
        return False

    # Get installation commands
    installation = get_installation_commands(stack_id, tier)

    logger.info("Installing npm dependencies...")
    logger.info("This may take several minutes...")

    runner = CommandRunner(default_timeout=600, working_dir=project_root)  # 10 min timeout

    # Map tier to command keys
    tier_to_command_key = {
        "tier-1-essential": ["base", "tier1"],
        "tier-2-standard": ["base", "tier1", "tier2"],
        "tier-3-comprehensive": ["base", "tier1", "tier2", "tier3"],
        "tier-4-production": ["base", "tier1", "tier2", "tier3", "tier4_dev", "tier4_prod"],
    }

    command_keys = tier_to_command_key.get(tier, ["base"])

    # Run installation commands sequentially
    for command_key in command_keys:
        if command_key not in installation["commands"]:
            logger.warning(f"Command key '{command_key}' not found for {stack_id}")
            continue

        command_str = installation["commands"][command_key]
        logger.info(f"Running: {command_key}...")

        # Parse command string and run
        result = runner.run(["sh", "-c", command_str], check=False)

        if not result.success:
            logger.error(f"Failed to install {command_key}: {result.stderr}")
            raise CommandExecutionError(
                command=command_str,
                returncode=result.returncode,
                stderr=result.stderr,
                context={"tier": tier, "command_key": command_key},
            )

        logger.info(f"✓ Installed {command_key}")

    logger.info("✅ All npm dependencies installed successfully")
    return True


def install_python_dependencies(
    tier: Literal[
        "tier-1-essential", "tier-2-standard", "tier-3-comprehensive", "tier-4-production"
    ],
    python_binary: str | None = None,
    project_root: Path | None = None,
) -> bool:
    """
    Install Python dependencies for ML/AI FastAPI stack.

    Args:
        tier: Quality tier
        python_binary: Python binary to use (e.g., "python3.11")
        project_root: Project root directory

    Returns:
        True if installation succeeded

    Raises:
        CommandExecutionError: If pip installation fails critically
    """
    if project_root is None:
        project_root = Path.cwd()

    if python_binary is None:
        python_binary = sys.executable

    # Get installation commands
    installation = get_installation_commands("ml_ai_fastapi", tier)

    logger.info("Setting up Python virtual environment...")

    runner = CommandRunner(default_timeout=600, working_dir=project_root)

    # Create virtual environment
    venv_path = project_root / "venv"
    if not venv_path.exists():
        result = runner.run([python_binary, "-m", "venv", "venv"], check=False)
        if not result.success:
            raise CommandExecutionError(
                command=f"{python_binary} -m venv venv",
                returncode=result.returncode,
                stderr=result.stderr,
            )
        logger.info("✓ Created virtual environment")
    else:
        logger.info("Virtual environment already exists")

    # Determine pip path
    pip_path = venv_path / "bin" / "pip"
    if not pip_path.exists():
        # Windows
        pip_path = venv_path / "Scripts" / "pip"

    # Upgrade pip
    logger.info("Upgrading pip...")
    result = runner.run([str(pip_path), "install", "--upgrade", "pip"], check=False)
    if result.success:
        logger.info("✓ Upgraded pip")

    # Install dependencies
    logger.info("Installing Python dependencies...")
    logger.info("This may take several minutes...")

    # Map tier to command keys (incremental installation)
    tier_to_command_key = {
        "tier-1-essential": ["base", "tier1"],
        "tier-2-standard": ["base", "tier1", "tier2"],
        "tier-3-comprehensive": ["base", "tier1", "tier2", "tier3"],
        "tier-4-production": ["base", "tier1", "tier2", "tier3", "tier4_dev", "tier4_prod"],
    }

    command_keys = tier_to_command_key.get(tier, ["base"])

    for command_key in command_keys:
        if command_key not in installation["commands"]:
            logger.warning(f"Command key '{command_key}' not found for ml_ai_fastapi")
            continue

        command_str = installation["commands"][command_key]
        # Replace 'pip install' with full pip path
        command_str = command_str.replace("pip install", f"{pip_path} install")

        logger.info(f"Running: {command_key}...")

        result = runner.run(["sh", "-c", command_str], check=False)

        if not result.success:
            logger.error(f"Failed to install {command_key}: {result.stderr}")
            raise CommandExecutionError(
                command=command_str,
                returncode=result.returncode,
                stderr=result.stderr,
                context={"tier": tier, "command_key": command_key},
            )

        logger.info(f"✓ Installed {command_key}")

    # Run security fixes if available
    if "security_fixes" in installation["commands"]:
        logger.info("Applying security fixes...")
        command_str = installation["commands"]["security_fixes"]
        command_str = command_str.replace("pip install", f"{pip_path} install")

        result = runner.run(["sh", "-c", command_str], check=False)
        if result.success:
            logger.info("✓ Applied security fixes")
        else:
            logger.warning("Security fixes failed (non-critical)")

    logger.info("✅ All Python dependencies installed successfully")
    return True


def install_dependencies(
    template_id: str,
    tier: Literal[
        "tier-1-essential", "tier-2-standard", "tier-3-comprehensive", "tier-4-production"
    ],
    python_binary: str | None = None,
    project_root: Path | None = None,
) -> bool:
    """
    Install dependencies for a template based on its package manager.

    Args:
        template_id: Template identifier
        tier: Quality tier
        python_binary: Python binary for Python projects
        project_root: Project root directory

    Returns:
        True if installation succeeded

    Raises:
        CommandExecutionError: If installation fails critically
    """
    # Determine package manager from template ID
    if template_id == "ml_ai_fastapi":
        return install_python_dependencies(tier, python_binary, project_root)
    else:
        return install_npm_dependencies(template_id, tier, project_root)
