requirements.txt
# Solokit (Session-Driven Development) - Python Dependencies
# Generated: 2025-10-18

# ============================================================================
# CORE DEPENDENCIES (Required for basic functionality)
# ============================================================================
# YAML parsing for API contract validation
PyYAML==6.0.1

# Configuration validation
jsonschema==4.20.0


# ============================================================================
# TESTING DEPENDENCIES (Required to run test suite)
# ============================================================================
pytest==8.2.2
pytest-cov==7.0.0


# ============================================================================
# CODE QUALITY TOOLS (Required for quality gates)
# ============================================================================
# Python linting and formatting
ruff==0.14.0

# Security scanning
bandit==1.7.10
safety==3.2.8


# ============================================================================
# OPTIONAL DEPENDENCIES (For specific features)
# ============================================================================
# Graph visualization (requires system graphviz installation)
# Uncomment if using /work-graph --format svg
# graphviz==0.20.1

# Docker support (for integration testing)
# docker==7.1.0
# docker-compose not available via pip (use system installation)


# ============================================================================
# DEVELOPMENT DEPENDENCIES (Optional)
# ============================================================================
# Uncomment for enhanced development experience:
# ipython==8.12.0  # Better REPL for debugging
